@@ -1,16 +1,29 @@
 module github.com/kscarlett/june
 
-go 1.19
+go 1.23
+
+toolchain go1.23.2
 
 require (
-	github.com/alecthomas/kong v0.7.1
-	github.com/yuin/goldmark v1.5.3
+	github.com/alecthomas/kong v1.11.0
+	github.com/yuin/goldmark v1.7.12
+)
+
+require (
+	github.com/BurntSushi/toml v1.2.1 // indirect
+	github.com/bep/godartsass/v2 v2.5.0 // indirect
+	github.com/bep/golibsass v1.2.0 // indirect
+	golang.org/x/sys v0.21.0 // indirect
+	google.golang.org/protobuf v1.35.2 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 require (
 	github.com/aymerick/douceur v0.2.0 // indirect
-	github.com/gorilla/css v1.0.0 // indirect
-	github.com/microcosm-cc/bluemonday v1.0.21 // indirect
+	github.com/fsnotify/fsnotify v1.9.0
+	github.com/gorilla/css v1.0.1 // indirect
+	github.com/microcosm-cc/bluemonday v1.0.27 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
-	golang.org/x/net v0.0.0-20221002022538-bcab6841153b // indirect
+	go.abhg.dev/goldmark/frontmatter v0.2.0
+	golang.org/x/net v0.26.0 // indirect
 )
